Don't export lvars from `or`, `not`, `not-join` clauses
It should generally be safe to assume that any new lvars introduced in a
`not` or `not-join` query clause are "negative": we're specifically trying to
ensure that they *can't* be bound, and therefore we should never add them to
the :find part of our generated queries. This also seems to be the case in
`or` clauses, but I'm not totally sure about this: it seems more plausible that
an `or` clause might genuinely introduce a new "positive" lvar that doesn't
appear elsewhere in the query. However, I don't have any examples of this
so far, and I *do* have examples of negative lvars introduced in `or` clauses
causing problems when exported, so I'm treating `or` clauses the same as
`not` and `not-join` clauses for now.